add a usecase package to organize application services
