+++ cmd/serve/serve.go
@@ -4,14 +4,14 @@ import (
 	"context"
 	e "errors"
 	"fmt"
+	relationshipsctl "github.com/project-kessel/inventory-api/internal/biz/usecases/relationships"
+	resourcesctl "github.com/project-kessel/inventory-api/internal/biz/usecases/resources"
 	"os"
 	"os/signal"
 	"syscall"
 	"time"
 
 	"github.com/project-kessel/inventory-api/cmd/common"
-	relationshipsctl "github.com/project-kessel/inventory-api/internal/biz/relationships"
-	resourcesctl "github.com/project-kessel/inventory-api/internal/biz/resources"
 	"github.com/project-kessel/inventory-api/internal/consistency"
 	"github.com/project-kessel/inventory-api/internal/consumer"
 	inventoryResourcesRepo "github.com/project-kessel/inventory-api/internal/data/inventoryresources"
