+++ internal/service/resources/hosts/hosts.go
@@ -2,11 +2,11 @@ package hosts
 
 import (
 	"context"
+	"github.com/project-kessel/inventory-api/internal/biz/usecases/resources"
 
 	pb "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/resources"
 	authnapi "github.com/project-kessel/inventory-api/internal/authn/api"
 	"github.com/project-kessel/inventory-api/internal/biz/model"
-	"github.com/project-kessel/inventory-api/internal/biz/resources"
 	"github.com/project-kessel/inventory-api/internal/middleware"
 	conv "github.com/project-kessel/inventory-api/internal/service/common"
 )
