+++ internal/service/resources/kesselcheckservice.go
@@ -2,11 +2,11 @@ package resources
 
 import (
 	"context"
+	"github.com/project-kessel/inventory-api/internal/biz/usecases/resources"
 
 	pb "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/authz"
 	authnapi "github.com/project-kessel/inventory-api/internal/authn/api"
 	"github.com/project-kessel/inventory-api/internal/biz/model"
-	"github.com/project-kessel/inventory-api/internal/biz/resources"
 	"github.com/project-kessel/inventory-api/internal/middleware"
 	"github.com/project-kessel/relations-api/api/kessel/relations/v1beta1"
 )
