+++ internal/service/resources/notificationsintegrations/notificationsintegrations.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/project-kessel/inventory-api/internal/biz/usecases/resources"
 	"io"
 	"time"
 
@@ -11,7 +12,6 @@ import (
 
 	"github.com/go-kratos/kratos/v2/log"
 	"github.com/project-kessel/inventory-api/internal/biz/model"
-	"github.com/project-kessel/inventory-api/internal/biz/resources"
 	"github.com/project-kessel/relations-api/api/kessel/relations/v1beta1"
 	"google.golang.org/protobuf/types/known/timestamppb"
